Make this works with angular, first dynamic version.
We are able to load static files from the static folder.
A new angular directive has been created in order to build the ace editor and load the content.
A new angular service has been created to load the static file content, and to return the loading promise.
Add a CSS file, in order to have a better code separation.